App: rename files (remove -new).
